% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_codepoint.R
\name{as_codepoint}
\alias{as_codepoint}
\title{Covert non-ASCII values to escaped codepoints}
\usage{
as_codepoint(x)
}
\arguments{
\item{x}{Value to convert.}
}
\value{
The escaped value.
}
\description{
Small convenience wrapper around iconv to help when you get the
'Portable packages must use only ASCII characters in their R code' message
from R CMD check.
}
\examples{

as_codepoint("£")

}
