% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_ages_to_interval.R
\name{imp_ages_to_interval}
\alias{imp_ages_to_interval}
\title{Fast categorisation of ages}
\usage{
imp_ages_to_interval(ages, limits = c(1L, 5L, 15L, 25L, 45L, 65L))
}
\arguments{
\item{ages}{\verb{[numeric]}.
Age in years. Values will be coerced to integer prior to categorising.}

\item{limits}{\verb{[numeric]}.
1 or more unique and positive cut points. Values will be coerced to integer.}
}
\value{
A factor of corresponding categories.
}
\description{
\code{imp_ages_to_interval()} provides fast categorisation of ages based on
specified limits. The resultant groupings will span the natural numbers
(from 0) and will always be closed on the left and open on the right. For
example, if \code{limits = c(1,10,30)} the groupings will be "[0, 1)",
"[1, 10)", "[10, 30)" and "[30, Inf)". This is comparable to a call of
\code{cut(ages, right = FALSE, breaks = c(0, limits, Inf))} but faster due to
the reduced flexibility of the function.
}
\examples{
imp_ages_to_interval(ages = 0:9, limits = c(3L, 5L))

}
