% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{imp_palette}
\alias{imp_palette}
\title{Plot a colour palette}
\usage{
imp_palette(values, label = TRUE, square = FALSE)
}
\arguments{
\item{values}{character vector of named or hex colours.}

\item{label}{boolean. Do you want to label the plot or not? If \code{values} is
a named vector the names are used for labels, otherwise, the values.}

\item{square}{boolean. Display palette as square?}
}
\value{
The input \code{values} invisibly.
}
\description{
\code{imp_palette()} plots a palette from a vector of colour values (name or hex).
}
\examples{
imp_palette(c("#5FE756", "red", "black"))
imp_palette(c("#5FE756", "red", "black"), square=TRUE)

}
\references{
Code inspired by https://stackoverflow.com/a/25726442
(h/t hrbrmstr) with the algorithm for text contrast coming from
https://stackoverflow.com/a/3943023.
}
