% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_name.R
\name{new_name}
\alias{new_name}
\title{Generate column names for a data frame}
\usage{
new_name(x, n = 1L)
}
\arguments{
\item{x}{A data frame.}

\item{n}{Number of unique names to generate.}
}
\value{
A character vector of unique names not already found in \code{x}.
}
\description{
\code{new_name()} generates unique names for additional data frame variables
ensuring they are not already present.
}
\examples{

new_name(mtcars)
new_name(mtcars, 2)

}
