% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot a colour palette}
\usage{
plot_palette(values, label = TRUE, square = FALSE)
}
\arguments{
\item{values}{character vector of named or hex colours.}

\item{label}{boolean. Do you want to label the plot or not? If \code{values} is
a named vector the names are used for labels, otherwise, the values.}

\item{square}{boolean. Display palette as square?}
}
\value{
The input (invisibly).
}
\description{
\code{plot_palette()} plots a palette from a vector of colour values (name or hex).
}
\examples{
plot_palette(c("#5FE756", "red", "black"))
plot_palette(c("#5FE756", "red", "black"), square=TRUE)

}
