% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{ympes-defunct}
\alias{ympes-defunct}
\alias{imp_assert_integer}
\alias{imp_assert_int}
\alias{imp_assert_double}
\alias{imp_assert_dbl}
\alias{imp_assert_numeric}
\alias{imp_assert_num}
\alias{imp_assert_logical}
\alias{imp_assert_lgl}
\alias{imp_assert_character}
\alias{imp_assert_chr}
\alias{imp_assert_data_frame}
\alias{imp_assert_list}
\alias{imp_assert_scalar_integer}
\alias{imp_assert_scalar_int}
\alias{imp_assert_scalar_double}
\alias{imp_assert_scalar_dbl}
\alias{imp_assert_scalar_numeric}
\alias{imp_assert_scalar_num}
\alias{imp_assert_scalar_logical}
\alias{imp_assert_scalar_lgl}
\alias{imp_assert_bool}
\alias{imp_assert_boolean}
\alias{imp_assert_scalar_character}
\alias{imp_assert_scalar_chr}
\alias{imp_assert_string}
\alias{ages_to_interval}
\title{Functions now defunct in package ympes}
\usage{
imp_assert_integer(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_int(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_double(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_dbl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_numeric(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_num(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_logical(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_lgl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_character(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_chr(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_data_frame(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_list(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_integer(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_int(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_double(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_dbl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_numeric(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_num(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_logical(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_lgl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_bool(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_boolean(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_character(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_chr(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_string(x, arg = deparse(substitute(x)), call = sys.call(-1L))

ages_to_interval(ages, limits = c(1L, 5L, 15L, 25L, 45L, 65L))
}
\description{
These functions are now defunct having previously been deprecated in version
0.3.0 of ympes. For replacement assertion functions see \code{help("assersions")}.
For a replacement to the \code{ages_to_interval()} function see
\code{ageutils::cut_ages()}.
}
\keyword{internal}
