% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ympes-deprecated}
\alias{ympes-deprecated}
\title{Deprecated functions in package \code{ympes}}
\usage{
breaks_to_interval(breaks)

cut_ages(ages, breaks)

split_interval_counts(
    lower_bounds,
    upper_bounds,
    counts,
    max_upper = 100L,
    weights = NULL
)

aggregate_age_counts(counts, ages = 0:(length(counts) - 1L), breaks)

reaggregate_interval_counts(
    lower_bounds,
    upper_bounds,
    counts,
    breaks,
    max_upper = 100L,
    weights = NULL
)
}
\description{
These functions are now deprecated and may be defunct as soon as the next
release of ympes.
}
\details{
Documentation for deprecated functions can be found via
\code{help("ageutils-deprecated")}.
}
\keyword{internal}
