% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ympes-package.R
\docType{package}
\name{ympes-package}
\alias{ympes}
\alias{ympes-package}
\title{ympes: Collection of Helper Functions}
\description{
Provides a collection of lightweight helper functions (imps) both for interactive use and for inclusion within other packages. These include minimal assertion functions with a focus on informative error messaging for both missing and incorrect function arguments as well as other functions for visualising colour palettes, quoting user input and searching rows of a data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TimTaylor/ympes}
  \item Report bugs at \url{https://github.com/TimTaylor/ympes/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [copyright holder]

}
\keyword{internal}
