% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Quote names}
\usage{
cc(..., .clip = FALSE)
}
\arguments{
\item{...}{Either unquoted names (separated by commas) that you wish to quote or a
length one character vector you wish to split by whitespace.

Empty arguments (e.g. third item in \verb{one,two,,four}) will be returned as \code{""}.

Character vectors not of length one are returned as is.}

\item{.clip}{\verb{[bool]}

Not currently used.}
}
\value{
A character vector of the quoted input.
}
\description{
\code{cc()} quotes comma separated names whilst trimming outer whitespace. It is
intended for interactive use only.
}
\examples{

cc(dale, audrey, laura, hawk)
cc("dale audrey laura hawk")

}
