% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Quote names}
\usage{
cc(..., .clip = getOption("imp.clipboard", FALSE))
}
\arguments{
\item{...}{Unquoted names (separated by commas) that you wish to quote.

Empty arguments (e.g. third item in \verb{one,two,,four}) will be returned as \code{""}.}

\item{.clip}{\verb{[bool]}

Should the code to generate the constructed character vector be copied to
your system clipboard.

Defaults to FALSE unless the option "imp.clipboard" is set to TRUE.

Note that copying to clipboard requires the availability of package
\href{https://cran.r-project.org/package=clipr}{clipr}.}
}
\value{
A character vector of the quoted input.
}
\description{
\code{cc()} quotes comma separated names whilst trimming outer whitespace. It is
intended for interactive use only.
}
\examples{

cc(dale, audrey, laura, hawk)

}
