% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{ympes-defunct}
\alias{ympes-defunct}
\alias{imp_assert_integer}
\alias{imp_assert_int}
\alias{imp_assert_double}
\alias{imp_assert_dbl}
\alias{imp_assert_numeric}
\alias{imp_assert_num}
\alias{imp_assert_logical}
\alias{imp_assert_lgl}
\alias{imp_assert_character}
\alias{imp_assert_chr}
\alias{imp_assert_data_frame}
\alias{imp_assert_list}
\alias{imp_assert_scalar_integer}
\alias{imp_assert_scalar_int}
\alias{imp_assert_scalar_double}
\alias{imp_assert_scalar_dbl}
\alias{imp_assert_scalar_numeric}
\alias{imp_assert_scalar_num}
\alias{imp_assert_scalar_logical}
\alias{imp_assert_scalar_lgl}
\alias{imp_assert_bool}
\alias{imp_assert_boolean}
\alias{imp_assert_scalar_character}
\alias{imp_assert_scalar_chr}
\alias{imp_assert_string}
\alias{ages_to_interval}
\alias{breaks_to_interval}
\alias{cut_ages}
\alias{split_interval_counts}
\alias{aggregate_age_counts}
\alias{reaggregate_interval_counts}
\title{Functions now defunct in package ympes}
\usage{
imp_assert_integer(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_int(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_double(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_dbl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_numeric(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_num(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_logical(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_lgl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_character(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_chr(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_data_frame(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_list(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_integer(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_int(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_double(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_dbl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_numeric(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_num(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_logical(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_lgl(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_bool(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_boolean(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_scalar_character(
  x,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)

imp_assert_scalar_chr(x, arg = deparse(substitute(x)), call = sys.call(-1L))

imp_assert_string(x, arg = deparse(substitute(x)), call = sys.call(-1L))

ages_to_interval(ages, limits = c(1L, 5L, 15L, 25L, 45L, 65L))

breaks_to_interval(breaks)

cut_ages(ages, breaks)

split_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  max_upper = 100L,
  weights = NULL
)

aggregate_age_counts(counts, ages = 0:(length(counts) - 1L), breaks)

reaggregate_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  breaks,
  max_upper = 100L,
  weights = NULL
)
}
\description{
These functions are now defunct.

For replacement assertion functions see \code{help("assertions")}.

For replacements to the age utilities function see the
\href{https://cran.r-project.org/package=ageutils}{\code{ageutils}} package.
}
\keyword{internal}
