% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ympes-package.R
\docType{package}
\name{ympes-package}
\alias{ympes}
\alias{ympes-package}
\title{ympes: Collection of Helper Functions}
\description{
Provides a collection of lightweight helper functions (imps) both for interactive use and for inclusion within other packages. These include functions for minimal input assertions, visualising colour palettes, quoting user input, searching rows of a data frame and capturing string tokens.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sr.ht/~tim-taylor/ympes/}
  \item \url{https://git.sr.ht/~tim-taylor/ympes}
  \item Report bugs at \url{https://lists.sr.ht/~tim-taylor/ympes}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item R Core Team (fstrcapture uses code from strcapture) [copyright holder]
  \item Toby Hocking (fstrcapture uses code from nc::capture_first_vec) [copyright holder]
}

}
\keyword{internal}
